fix: change the project's folder structure and modify skill model
