+++ models.py
@@ -2,10 +2,6 @@ from datetime import date
 from typing import Optional
 from pydantic import BaseModel, EmailStr, Field
 
-class Skill(BaseModel):
-    name : str = Field(...)
-    type : str = Field(...)
-
 class Experience(BaseModel):
     title : str = Field(...) 
     company : str = Field(...)
